Refactor into util files
